reorganise demo code
