+++ demo/aliasing.rs
@@ -1,5 +1,5 @@
-// Kani: kani demo/darpa/aliasing.rs --output-format terse
-// Soteria Rust: dune exec -- soteria-rust exec demo/darpa/aliasing.rs --kani
+// Kani: kani demo/aliasing.rs --output-format terse
+// Soteria Rust: dune exec -- soteria-rust exec demo/aliasing.rs --kani
 
 /// An example of Tree Borrows violation, according to the paper.
 /// We may only own one mutable reference to a memory location at a time;
